Pull access-to-subprogram types contracts into proof dependencies
This patch adds proof dependencies on the contracts of an access-to-subprogram type every time the type is encountered, like what is done for predicates.

[changelog]
* src/flow/flow_utility.adb
* src/flow/flow_utility.ads
* src/flow/flow_utility-proof_dependencies.adb
* src/flow/flow_utility-proof_dependencies.ads
(Process_Type_Contracts, Process_Type_Contracts_Internal): Renamings
of Process_Predicate_And_Invariant(_Internal). In addition to predicates
and invariants, pull contracts of access-to-subprogram types in proof
dependencies.