Add conversion to wrapper type in is_moved property for records
It can happen that a wrapper module exists for the type of a record
field even if it does not exist for the record type itself and the
type of the field is not annotated with Relaxed_Initialization (because
tagged types never have initialization wrappers).
In this case, it is necessary to convert the field access to the
initialization wrapper before calling _is_moved_or_reclaimed
recursively on the component.