Mandate precondition when potentially unsound for overridden dispatching primitives
When the public view of a type is untagged, the contract for overridden dispatching primitive is chosen differently depending on whether the call is regular or dispatching. If there is no specific precondition, this can cause
unsoundness. A similar phenomenon can happen with default selection of
class-wide precondition due to change of visibility caused by hiding
private parts. Avoid the issue by mandating a class-wide precondition in
these cases.