Pick variables used in abnormal raise statements
With the new handling of non-returning subprograms we must pick
variables used in the string_expression of raise statements.

Likewise, it seems better to pick subprogram calls and proof
dependencies from there.

* src/flow/flow-control_flow_graph.adb
(Do_Raise_Statement): Process expression not just for a goto-like
exception, but also for an abnormally terminating raise.
* testsuite/gnatprove/tests/191__flow_raise/
Update output, which for a raise statement is now similar to the output
for a raise expression.