Support for Program_Exit in proof
Like for exceptions, we mandate a Program_Exit contract as soon as a
subprogram might abrupty termiante the program. It provides a boolean
condition that should evaluate to True when the program is terminated.
This condition is only allowed to mention values of the subprogram in
the pre-state and values of standalone objects in the post-state.
Outputs of the subprogram that are not mentioned in the program exit
condition do not need to be left in a valid state (they might be
uniniliazed, unreclaimed, or even aliased (moved or borrowed)). Those
mentioned in the condition shall be readable.
On calls to subprograms annotated with Program_Exit, checks are emitted
to enforce the program exit condition of the subprogram if any, or that
the program cannot be exited otherwise.
The Exit_Cases aspect is extended to support a new Program_Exit exit
kind.