Support Exit_Cases contracts
There are several ways for a subprogram to terminate: it can return
normally or propagate an exception. If a subprogram does not always
terminate normally, then it is possible to annotate it with an
Exit_Cases aspect. It allows partioning the input state into cases,
specifying for each case what the expected termination kind is. It can
be either Normal_Return, if the subprogram shall return normally,
Exception_Raised, if it shall raise an (unspecified) exception, or
(Exception_Raised => E), if it shall raise the exception E.