On raise handle 'Result just like a formal parameter of mode OUT
In general, flow analysis represents 'Result like a formal parameter of mode
OUT. For consistency, we now also "scrub" it when an exception announced in
Exceptional_Cases is raising by a function with side effects. This removes an
otherwise confusing check.

[changelog]
* src/flow/flow-control_flow_graph.adb
(Create): Add 'Result to variables "scrubbed" when raising an exception.
* testsuite/gnatprove/tests/359__procedural_functions/test.out
* testsuite/gnatprove/tests/408__aliasing/test.out
* testsuite/gnatprove/tests/645__side_effects_decl_illegal/test.out
* testsuite/gnatprove/tests/645__side_effects_decl_legal/test.out
* testsuite/gnatprove/tests/783__exception_relaxed_result/test.out
* testsuite/gnatprove/tests/786__function_with_exception/test.out
* testsuite/gnatprove/tests/791__flow_side_effects_raise/test.out
Unnecessary checks are gone.