Fix command line usage of gnatcov
A previous change tried to make it possible to instrument both gnat2why
and gnat2why_gnat for coverage, but this apparently hadn't been tested:
the --projects switch of gnatcov was used incorrectly in that case.

The incorrect code assumed that --projects accepted a comma-separated
list of project names, while it actually only accepts one name but can
be passed multiple times.

Note: instrumenting the two projects still doesn't work with this change
for other reasons.