Support membership tests with different base types
In membership tests, the tested expression might be of a scalar type
which does not have the same base as the types used in membership
choices (in particular if the expression has the universal integer
type). In this case, comparisons should be done in a type compatible
both with the tested expression and the membership choice.