Info messages become warnings disabled by default
Most info messages become warnings that are disabled by default, and can
be enabled individually via the -W switch, or globally using `--info`.
This patch adds the warning kinds, tags, messages and descriptions for
all these new warnings.

The only exception are loop unrolling and inlining messages, which stay
info messages and are now issued by default. They can be disabled with
"-A info-unrolling-inlining".

Previously, the implementation of the `--info` switch used the front-end
debug flag Debug_Flag_Underscore_F. This flag is still used, but only
for the front-end messages related to unrolling and inlining. If the
flag is true, these messages are issued. The flag is set to false if the
users passes the "-A info-unrolling-inlining" switch. The gnat2why
back-end doesn't use this flag anymore.

Some more changes:
- The UG now mentions all those warning tags.
- It's not needed to pass the "--info" switch to gnat2why, the
  information is contained in the warning status array
- "--info" is now a file-specific switch
- The code for the info messages now uses Warning_Msg_N. One warning now
  uses a continuation message.
- "ug__" tests now use "-A info-unrolling-inlining" to not have those
  messages in the UG output.
- One long info message is split into a warning + continuation message