Restore and tune check for effectively volatile function
While functions with volatile parameter or result type but no Volatile_Function
aspect are rejected in marking, we still need to recognize them as genuinely
volatile functions in flow. Otherwise we will emit a spurious warning saying
that they are declared as volatile but have no volatile effect.

This patch fixes a recent regression, but also removes a spurious warning
on functions that were volatile because of their result type was volatile.

[changelog]
* src/flow/flow-analysis.adb
(Check_Function_For_Volatile_Effects): Check both parameters and result for
volatile types.
* testsuite/gnatprove/tests/819__flow_volatile_result/test.out
Spurious warning is gone.