No longer recognize runtime alternatives for SPARKlib units
The units Ada.Numerics.Big_Numbers.Big_Integers_Ghost and
System.SPARK.Cut_Operations used to be alternatives for SPARKlib
units SPARK.Big_Integers (on light runtimes) and SPARK.Cut_Operations.
They were used to verify runtime units with SPARK, that could not
depend on the SPARKlib. Now that this effort is being stripped out of
the runtime, these units are not needed anymore and are removed as
well. Remove their support in SPARK.