Allow named access-to-constant as observing parameter
It is now possible to pass an object of a named access-to-constant
type as a parameter to a subprogram call expecting an anonymous
access-to-constant type. The rules are not relaxed for other
kinds of observes, in particular, when declaring or assigning into
a standalone object of an anonymous access-to-constant type.