Fix excessive memory consumption for records with many subcomponents
For declaration and assignment of record objects, we maintained a map from
every vertex representing this declaration/assignment to all the other vertices
coming from the same declaration/assignment. This required a quadratic amount
of memory wrt the number of components in the flattened view.

This patch replaces the map with a pointer to the first vertex in the sequence
that represents the declaration/assignment. The other vertices are then picked
from the CFG structure itself.

[changelog]
* src/flow/flow-analysis.adb
(Other_Field_Is_Effective): Pick vertices using new attribute and the CFG
itself.
* src/flow/flow-control_flow_graph.adb
(Do_Assignment_Statement, Do_Object_Declaration): Set new attribute in the
sequence of vertices; also, set the cluster for graph printing in both phase 1
and 2 (it was an oversight to only set it in phase 2).
* src/flow/flow.ads (Flow_Analysis_Graphs_Root): Remove component responsible
for excessive memory consumption.
* src/flow/flow_types.ads (V_Attributes, Null_Attributes): New field.