Forbid propagation of exceptions from finally statements
We forbid exceptions propagating from the body of finally statements
in SPARK. This simplifies control flow analysis by allowing only one
pattern of exit for finally statements. Such a restriction is in line
with the intuition that finalization code should not affect control
flow. This also matches the existing restrictions of GNAT on static
transfer-of-control statements (goto/return/exit) within finally
statements.