Change entity reference in flow warning reporting
Flow would crash when emitting a message about unwritten exports
when the entity was of access type whose designated type is a
completion of an incomplete type (because the complete view is
marked as internal). This patch modifies the entity in the message
so that the crash no longer appears.