Take into account difference between top/deep levels in borrow-checker
When performing an operation (read/write/observe/borrow/etc), the
permission requirement may differ between components at top level
(shallow components), and components that may be after dereferences.
For example, when making a move, the toplevel components can be read-only,
since they are copied. However, the components below dereferences are
moved, so we need full permission to them. A similar scenario occur for
borrow.

This is relevant for proper treatment of 'Access: values obtained
by 'Access are read-only, as they are not even stored at an affectable
location. However, the dereferenced values may have different permissions.