Use static value of attribute Constrained prefixed with variables
For GNATprove, the value of 'Constrained prefixed with variables is statically
known to the frontend. It is only in expansion of protected entries and accept
statements when their formal parameters are rewritten into constant and
variable entities whose 'Constrained must be computed at run time.

[changelog]
* src/spark/spark_util.adb (Attr_Constrained_Statically_Known): When prefix
is variable, then the attribute is known statically.
* testsuite/gnatprove/tests/N617-011__mutable_discr/test.out
* testsuite/gnatprove/tests/O120-034__constrained/test.out
* testsuite/gnatprove/tests/O403-005__private_attributes/test.out
* testsuite/gnatprove/tests/S711-051__constrained_attr/test.out
Update outputs.