Fix warning about duplicated Get_Targ unit
GNATstudio, which uses gprbuild, was rightly warning about two source files
providing the Get_Targ unit: one was in the GNAT source directory, the other
was its copy renamed when setting up the gnat2why project.

Now we don't create this copy and use GPR to deal with a renaming.