Remove logic for implicit inputs from interprocedural graph
Logic for implicit inputs was duplicated between creating 'In vertices
(i.e. Make_Parameter_Attributes and Make_Global_Attributes) and building
the interprocedural calls. This was just a premature optimization to avoid data
dependency edges leading to 'In vertices that use no variables. It didn't
reduce computational complexity and was just making the code more difficult
to maintain.

Behavior of the tool is unaffected.

[changelog]
* src/flow/flow-interprocedural.adb
(Is_Implicit_Input): Remove; adapt callsite.