Allow move on more than one path for shallow objects
As a performance optimization for cases with many (shallow) objects/paths,
shallow objects were not tracked in the main permission environment, so that
the implementation would not have to handle too many objects during path merges.
Shallow objects were instead added in a dedicated environment upon moves,
environment which was only reset at subprogram start. As such, if they were moved
on independent paths, one of the moves would still be considered to prevent the
other.

To handle these objects correctly, we put them in the main permission
environment, duplicated with a starting value environment to keep most of
the benefits of the opt-in optimization. The principle is that elements
left in the starting value environment are trivial to merge. However, if
any non-trivial borrow-checking is requested for them, we move the value
within the main permission environment so that it is tracked adequately
afterward.