Pick refined globals when generating globals for bodies not in SPARK
When picking unrefined globals for subprograms whose bodies are not in SPARK,
we must also fill the refined globals, because otherwise the unrefined ones
will be erased when up-projecting the refined ones.

[changelog]
* src/flow/flow_generated_globals-partial.adb
(Down_Project): Lift so it can be reused.
(Preanalyze_Spec): Fill refined globals based on the proper ones.
* testsuite/gnatprove/tests/859__rover/test.opt
Enable; no longer crashes.