Add violations for relaxed initialization
As per the RM wording, objects and types annotated with
Relaxed_Initialization should not have parts of an unchecked union
type or be effectively volatile. The Relaxed_Initialization should not
be allowed on primitives of tagged types.

Also fix a known limitation that would lead to Program_Error being
raised when calling Contains_Relaxed_Init_Parts on partially marked
code by introduing a new Check_No_Relaxed_Init_Part procedure that
registers unmarked components types to be checked later.

Finally, raise a violation instead of a limitation on parts of tagged
types.