Support Potentially_Invalid in proof
For now, locally potentially invalid values are not supported.
Add a validity flag for objects annotated with Potentially_Invalid.
This flag is copied when inside a declaration, assignment, or store
for actual parameters if they propagate the validity status.
Otherwise, the flag is checked as soon as the object is evaluated
except as a prefix of the 'Valid or 'Valid_Scalar attributes, when
accessing array bounds or discriminants, and as the prefix of Size.