Emit an init check on top-level discriminants of actual parameters
On exceptional exit, emit an unprovable initialization check for the
mutable discriminants of parameters that are passed neither by copy nor
by reference if they are annotated with Relaxed_Initialization. It is
necessary as proof considers mutable discriminants of standalone objects or
parameters to always be initialized (there are not flag for them).

Add a way to store an explanation for a proof check when it is created.