Add test with a wrong location for inlined range check
Test output includes message about range check proved at the location of a call
to procedure P2, but this location is wrong. There is no range check there; the
range check occurs in the inlined call. This is a minimal reproducer extracted
from ACATS.

[changelog]
* testsuite/gnatprove/tests/933__inline_sloc/
Add test; the output includes a wrong location.