Add test with uninitialized volatile variable
It seems wrong to not emit any check for this code, since there is access to
uninitialized data from the main procedure.

For completeness, the original test is extended to cover not just volatile
uninitialized variables, but also volatile initialized variables and volatile
imported variables, which we all discussed when looking for a fix.