Move fixup for fixedpoint division and multiplication to expansion
An AST fixup for fixedpoint division and multiplication inside a type
conversion was done in rewriting, which was convenient, but caused a range
check to be missed. Now it is done by a minimal expansion of the converted
expression in the frontend, so that the range checks are added when processing
the enclosing conversion itself.

[changelog]
* src/spark/spark_rewrite.adb
(Rewrite_Fixed_Point_Mult_Div): Remove.
* testsuite/gnatprove/tests/939__fixedpoint_conv/test.out
* testsuite/gnatprove/tests/KB28-016__over/test.out
New checks in both a dedicated test and an old ACATS test appear.