Add test where generated global involves a higher-order function
This test checks that flow picks a higher-order actual with no explicit
Global/Depends contract when collecting subprograms called. This higher-order
actual then contributes its globals to the globals of the higher-order callee.

Code-wise, it checks that we include the higher-order actual in
Pick_Generated_Info, as otherwise we would crash on incomplete generated
globals.