Fix crash when explaining array indexed by loop parameter
Replace call to low-level relaxed initialization query (which is meant to crash
on ill-defined inputs) with a call to high-level utility.

[changelog]
* src/flow/flow_error_messages.adb (Explain_Variables): Change routine to ask
for relaxed initialization.
* testsuite/gnatprove/tests/970__relaxed_loop_param/
Add test; with this patch it no longer crashes.
* testsuite/gnatprove/tests/RC17-014__init_by_proof/test.out
* testsuite/gnatprove/tests/T212-013__relaxed_init_function_result/test.out
* testsuite/gnatprove/tests/V328-024__relaxed_init/test.out
* testsuite/gnatprove/tests/V809-008__relaxed_globals/test.out
Improve explanations for objects whose relaxed initialization come from type
and not from the explicit object aspect.