Support duplicated contracts
To support multiple levels of ghost code, the frontend is going to
introduce multiple pragmas for a single contract aspect. As a result,
GNATprove has to adapt to expect several pragmas when there could
previously only be one per aspect.

Pre and postconditions could already be duplicated. They are handled
through the Find_Contracts function that returns a list of expressions.
We now also use Find_Contracts for Initial_Condition and Contract_Cases.
In addition, the proof support for contract cases used to introduce an
object for the others case of the contract cases aspect of a subprogram,
the other cases being stored in a map. To support multiple contract
cases with an others case, the object has been replaced by an additional
mapping in the map.

Type contracts are handled through frontend generated subprograms. We
update the subprogram querying the type constraints from check only
procedures to return several expressions if there are several check
pragma. Locate checks for DIC on the DIC expression instead of on the
type. Currently, all applicable invariants are checked at once, so we
don't attempt to relocate them.
Using Get_Pragma is still OK to search for a DIC of null as it is not
allowed to use assertion levels in this case.

We reject duplication of Subprogram_Variant. It seems useless.