Update outputs with checks in prefixes of array attributes
Accept current outputs in non-dedicated tests.

[changlog]
* testsuite/gnatprove/tests/M318-008__flow_generics/test.out
Loop is no longer unrolled, because the prefix of
"Record.Variant_Static_Array'Range" doesn't statically name an object.
* testsuite/gnatprove/tests/O916-052__crazyflie/test.out
A missing check rightly appeared.
* testsuite/gnatprove/tests/O512-022__model_transform/test.out
Expression of the form "Function_Call (Args)'Range" is not expanded, because
it doesn't statically name an object, so consequently we get some range checks.
* testsuite/gnatprove/tests/P113-023__alias_issue/test.out
Illegal code; I didn't investigate further.
* testsuite/gnatprove/tests/R416-004__frame_condition/test.out
Expressions of the form "Slice (1 .. 3)'Range" are not expanded, because
it doesn't statically name an object, so consequently we get some range checks.
* testsuite/gnatprove/tests/TB10-015__inlining_others/test.out
Expression coming from inlining-from-proof is not expanded, because it involves
unchecked type conversion and doesn't statically name an object.