Minor optimizations for overlays of constant objects
Constant overlays are like UC from the overlaid objects to the
overlaying object. As the UC does not go the other way around, it is
not necessary to require that the overlaid object is suitable as the
source of an unchecked conversion. It can also safely be annotated with
Potentially_Invalid.