Make flow graph vertices unkeyed by default
Adding vertices to the control flow graph was also creating a unique mapping
from a Flow_Id (which was typically holding a Node_Id) to Vertex_Id. This was
providing an easy access from vertex to a corresponding Flow_Id/Node_Id, but
it was actually only used for 'Initial and 'Final vertices.

This unique mapping is incompatible with the upcoming support for finally
statements, where every execution of those statements will be represented on
the graph by a separate vertex, so every statement from the finally section may
be represented by multiple vertices.

This patch adapts the API of the flow graphs to make their vertices unkeyed by
default and the API of the underlying graphs to keep their vertices keyed by
default.

[changelog]
* src/flow/flow-control_flow_graph.adb
(Add_Vertex): Add Keyed parameter which defaults to False.
(Create_Initial_And_Final_Vertices): Create keyed vertices for objects.
(Create): Likewise for subprograms when deciding whether they are called
definitely or conditionally, which is part of the global generation process.
* src/flow/graphs.ads
(Add_Vertex): Add Keyed parameter which defaults to True; spec.
* src/flow/graphs.adb
(Add_Vertex): Likewise; body.