Refine ceiling priority check messages and add more tests
This patch extends the ceiling priority check messages by providing references
to the locked object, the target of the potentially offending call and a full
call trace from the source location associated with the check.

It also resolves the shortcoming when some ceiling priority violation were
not detected over a call chain: eng/spark/spark2014#1008.