Fix spurious warning with extended return statement
When creating a graph for an extended return statement we always created a
vertex where Function'Result is assigned. However, when statements inside the
extended return does not return normally, e.g. because there is a simple
return, then this vertex is useless (the simple return statement recognizes
when it is inside an extended return and creates a similar vertex itself).

This useless vertex causes a spurious warning saying that it is never reached.

This was not particularly problematic until now, but it needs to be fixed to
properly support finally statements, which similarly must be executed only once
and not both my extended and simple return statements.

[changelog]
* src/flow/flow-control_flow_graph.adb
(Do_Extended_Return_Statement): Only create vertex for implicit return when an
implicit return is actually executed.