Always print reference to the locked object in ceiling priority messages
This avoids producing confusing output when some ceiling priority checks at
a given callsite are proved and others not. By always providing references
to the checked objects it is clear which objects are concerned.

[changelog]
* src/flow/flow_error_messages.adb
(Error_Msg_Proof): Add an optional Subject section to the check messages.
* src/spark/spark_util.ads
(Check_Info_Type): Add field for the optional Subject section.
* src/why/gnat2why-subprograms.adb
(Detailed_Msg): Rename function to Msg_Details. Add the generation of the
   Subject section for ceiling priority messages.
(Create_Check): Add the generation of the Subject section for ceiling priority
   messages.
* src/why/gnat2why-util.ads
(New_Check_Info): Add optional parameter for the subject of the check.
* src/why/gnat2why-util.adb
(New_Check_Info): Same.
* testsuite/gnatprove/tests/112__ceiling_prio4_records_basic/*
   Extend test to produce multiple positive and negative messages.
* testsuite/gnatprove/tests/*.out
   Update baselines.