Register priority and max queue length only for called operations
For composite objects we registered priority for all components with protected
types and similarly we registered max queue length for all entries.

Now we register this information only for components and entries that are
actually accessed by subprograms analyzed in the current unit.

This removes unnecessary information written in phase 1, while phase 2 and the
overall behavior is unaffected.