Remove record tree when removing dead object declaration
When removing 'Initial and 'Final vertices of an object declared in what is
found to be a dead code, we now also remove its record tree. This makes CFGs
easier to inspect at; otherwise, semantics is unaffected.

[changelog]
* src/flow/flow-control_flow_graph.adb
(Create_Record_Tree): New parameter with the currently processed entity; keep
vertex identified after adding it, store it in the container and reuse when
linking to the graph (this could be done before, to avoid unnecessary search).
(Create_Initial_And_Final_Vertices): Pass currently processed entity, if
present.