Handle dynamically constrained arrays in overlays
Turn checks for matching sizes of overlays into dynamic checks on
arrays slices and dynamically constrained array types with aliased
components. These checks compute the dynamic size by multiplying
the number of elements by the component size. This is valid when
the components are aliased. Also use the component type when
checking whether an array with aliased components has holes.