Avoid introducing sliding on record component update
When updating a record component, it can happen that the expected it
and the effective type of the component do not match when the effective
type is a bounded Itype introduced by the frontend for a discriminant
dependent component. In this case, sliding is never necessary, and
trying to introduce it might cause the tool to crash as it encounters
discriminants for which there is no value in the context.