Use ultimate alias to handle overlays
In proof and borrow checking, no longer consider mutable overlays as
distinct objects. Instead, go to the ultimate alias. This is consistent
with what is already done in flow analysis.

For now, use the frontend routine Ultimate_Overlaid_Entity to get the
root object of an overlay. It is consitent with flow analysis but
probably not what we want in the long term as it might break SPARK
boudaries and does not match the definition of precisely supported
address clauses. In the meantime, raise an exception in proof code when
there is a mismatch.
The special handling for overlays in proof specific flow routines has
been removed, as now the handlings in proof and flow is consistent.
In the borrow checker, consider reads and writes to an overlay as reads
or writes to the root. Still reject moves, observes, and borrows of
overlaid objects as unsupported.
In proof, use a function for mutable symbols of overlaid objects. It
takes the root as a parameter. Writing an overlay havocs the root
object and assumes the new value of the overlaid object. The fact that
the overlaid objects are valid - that their dnamic invariant holds -
for all valid values of the root is only assumed in scopes where the
overlay is declared. For frame condition computation, add a post
processing that merges the write statuses of overlays.