Keep local array objects on the heap in graph algorithms
When doing a depth-first search and computing the dominator tree of a graph
we are using arrays. Since they have as many elements as the vertices in the
graph, on large graphs we were overflowing the stack. This patch moves those
arrays from stack to heap and secondary stack.

[changelog]
* src/flow/graphs.adb
(DFS): Move local array to heap.
(Dominator_Tree_Internal): Move local arrays to heap and the returned array
to secondary stack.